Add files via upload
This is a full guided project on SQL by Angelina Frimpong based on sales data-set. Here the main objective of the project was to explore a sales data-set and determine various insights such as sales by year, sales by product line, sales by deal size. Along with such insights , the best customer and the products that were sold together were discovered using RFM Analysis which is a marketing technique used to quantitatively rank and group customers based on the recency, frequency and monetary total of their recent transactions to identify the best customers and perform targeted marketing campaigns. 

The steps to complete the project were:
 Importing a file into SQL Server Database
- SQL Aggregate Functions
- SQL Window Functions
- SQL Sub Query
- Common Table Expressions (CTEs)
- SQL XML Path Function
- RFM Analysis